feat: vote and comment counters
