+++ labeddit/src/pages/PostPage/PostPage.js
@@ -2,15 +2,10 @@ import React, { useEffect, useState, useContext } from 'react';
 import axios from 'axios';
 import { useNavigate, useParams } from 'react-router-dom';
 import useProtectedPage from '../../hooks/useProtectedPage';
-import Header from '../../components/Header';
+import Header from '../../components/Header/Header';
 import { BASE_URL } from '../../constants/urls';
 import {
   Container,
-  Post,
-  SendBy,
-  BodyPost,
-  CountersDiv,
-  CountersContainer,
   PostComment,
   Comment,
   BtnButton,
@@ -18,12 +13,11 @@ import {
   CommentsContainer,
   CloseBtn
 } from './styled';
-import arrowUp from '../../assets/arrow_up.png';
-import arrowDown from '../../assets/arrow_down.png';
-import speechBubble from '../../assets/speech_bubble.png';
 import closeButton from '../../assets/close_button.png';
 import { GlobalContext } from '../../context/global/GlobalState';
 import { goBack } from '../../routes/coordinator';
+import PostCard from '../../components/PostCard/PostCard';
+import CommentCard from '../../components/CommentCard/CommentCard';
 
 function PostPage() {
   useProtectedPage();
@@ -58,7 +52,8 @@ function PostPage() {
         headers: { Authorization: token }
       })
       .then(response => {
-        setComments('');
+        setComments([]);
+        getPostComment();
       })
       .catch(error => console.log(error));
   };
@@ -66,7 +61,6 @@ function PostPage() {
   const handleChange = event => {
     setComments(event.target.value);
   };
-  useEffect(createComment, []);
 
   return (
     <Container>
@@ -76,21 +70,7 @@ function PostPage() {
         alt="close button"
         onClick={() => goBack(navigate)}
       />
-      <Post key={post.id}>
-        <SendBy>Enviado por: {post.username}</SendBy>
-        <BodyPost> {post.body} </BodyPost>
-        <CountersContainer>
-          <CountersDiv>
-            <img src={arrowUp} alt="arrow up" />
-            <p> {post.voteSum} </p>
-            <img src={arrowDown} alt="arrow down" />
-          </CountersDiv>
-          <CountersDiv>
-            <img src={speechBubble} alt="speech bubble" />
-            <p> {post.commentCount} </p>
-          </CountersDiv>
-        </CountersContainer>
-      </Post>
+      <PostCard key={post.id} post={post} />
       <PostComment>
         <Comment
           rows="4"
@@ -104,19 +84,8 @@ function PostPage() {
       <Divider />
       <CommentsContainer>
         {postComment.map(comment => {
-          console.log(comment);
           return (
-            <Post>
-              <SendBy>Enviado por: {comment.username}</SendBy>
-              <BodyPost> {comment.body} </BodyPost>
-              <CountersContainer>
-                <CountersDiv>
-                  <img src={arrowUp} />
-                  <p> {comment.voteSum} </p>
-                  <img src={arrowDown} />
-                </CountersDiv>
-              </CountersContainer>
-            </Post>
+            <CommentCard comment={comment}/>
           );
         })}
       </CommentsContainer>
