+++ labeddit/src/pages/PostPage/styled.js
@@ -14,45 +14,6 @@ export const CloseBtn = styled.img`
   margin: 13px 0 13px 34px;
 `;
 
-export const Post = styled.li`
-  width: 90%;
-  border-radius: 12px;
-  border: 1px solid #e0e0e0;
-  background: #fbfbfb;
-  list-style: none;
-  margin-top: 28px;
-  text-align: left;
-  padding: 10px;
-  box-sizing: border-box;
-  word-wrap: break-word;
-`;
-
-export const SendBy = styled.p`
-  font-size: 12px;
-  color: #6f6f6f;
-`;
-
-export const BodyPost = styled.p`
-  font-size: 18px;
-`;
-
-export const CountersContainer = styled.div`
-  display: flex;
-  width: 100%;
-`;
-
-export const CountersDiv = styled.div`
-  display: flex;
-  align-items: center;
-  border: 1px solid #e0e0e0;
-  border-radius: 12px;
-  padding: 3px 10px;
-  height: 20px;
-  font-size: 10px;
-  font-weight: 700;
-  margin-right: 10px;
-  gap: 10px;
-`;
 
 export const PostComment = styled.div`
   display: flex;
